Auth Specifications
Signup with email, password, and password confirmation.
Login with email and password.
Logout when logged in.
Change password with current and new password.
Signup and Sign in must only be available to not signed in users.
Logout and Change password must only be available to signed in users.
Give feedback to the user after each action's success or failure.
All forms must clear after submit success and user sign-out