Styling/Debugging
User is unable to click out of the add workout and change password
modals. They have to click the 'x' to close. This is to ensure that
the success/failure messages disappear when a user clicks out of the
form in the modal

The nav toggler now collapses when any of the three nav bar items are
clicked. This helps with a bug that was happening when a user signed
out, but when they signed back in the nav bar would not be collapsed.